Moved types around and test skeleton.
