+++ node/src/integration-test/kotlin/net/corda/node/NodeStatePersistenceTests.kt
@@ -1,37 +1,34 @@
-package net.corda.test.node
+package net.corda.node
 
 import co.paralleluniverse.fibers.Suspendable
+import net.corda.MESSAGE_CONTRACT_PROGRAM_ID
+import net.corda.Message
+import net.corda.MessageContract
+import net.corda.MessageState
 import net.corda.client.rpc.CordaRPCClient
-import net.corda.core.contracts.*
+import net.corda.core.contracts.Command
+import net.corda.core.contracts.StateAndContract
+import net.corda.core.contracts.StateAndRef
 import net.corda.core.flows.FinalityFlow
 import net.corda.core.flows.FlowLogic
 import net.corda.core.flows.StartableByRPC
-import net.corda.core.identity.AbstractParty
 import net.corda.core.identity.Party
+import net.corda.core.internal.packageName
 import net.corda.core.messaging.startFlow
-import net.corda.core.schemas.MappedSchema
-import net.corda.core.schemas.PersistentState
-import net.corda.core.schemas.QueryableState
-import net.corda.core.serialization.CordaSerializable
-import net.corda.core.transactions.LedgerTransaction
 import net.corda.core.transactions.SignedTransaction
 import net.corda.core.transactions.TransactionBuilder
 import net.corda.core.utilities.ProgressTracker
 import net.corda.core.utilities.getOrThrow
 import net.corda.node.services.Permissions.Companion.invokeRpc
 import net.corda.node.services.Permissions.Companion.startFlow
-import net.corda.testing.node.User
 import net.corda.testing.core.singleIdentity
 import net.corda.testing.driver.DriverParameters
-import net.corda.testing.driver.PortAllocation
 import net.corda.testing.driver.driver
 import net.corda.testing.driver.internal.RandomFree
+import net.corda.testing.node.User
 import org.junit.Assume.assumeFalse
 import org.junit.Test
 import java.lang.management.ManagementFactory
-import javax.persistence.Column
-import javax.persistence.Entity
-import javax.persistence.Table
 import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 
@@ -44,7 +41,7 @@ class NodeStatePersistenceTests {
 
         val user = User("mark", "dadada", setOf(startFlow<SendMessageFlow>(), invokeRpc("vaultQuery")))
         val message = Message("Hello world!")
-        val stateAndRef: StateAndRef<MessageState>? = driver(DriverParameters(isDebug = true, startNodesInProcess = isQuasarAgentSpecified(), portAllocation = RandomFree)) {
+        val stateAndRef: StateAndRef<MessageState>? = driver(DriverParameters(isDebug = true, startNodesInProcess = isQuasarAgentSpecified(), portAllocation = RandomFree, extraCordappPackagesToScan = listOf(MessageState::class.packageName))) {
             val nodeName = {
                 val nodeHandle = startNode(rpcUsers = listOf(user)).getOrThrow()
                 val nodeName = nodeHandle.nodeInfo.singleIdentity().name
@@ -78,7 +75,7 @@ class NodeStatePersistenceTests {
 
         val user = User("mark", "dadada", setOf(startFlow<SendMessageFlow>(), invokeRpc("vaultQuery")))
         val message = Message("Hello world!")
-        val stateAndRef: StateAndRef<MessageState>? = driver(DriverParameters(isDebug = true, startNodesInProcess = isQuasarAgentSpecified(), portAllocation = RandomFree)) {
+        val stateAndRef: StateAndRef<MessageState>? = driver(DriverParameters(isDebug = true, startNodesInProcess = isQuasarAgentSpecified(), portAllocation = RandomFree, extraCordappPackagesToScan = listOf(MessageState::class.packageName))) {
             val nodeName = {
                 val nodeHandle = startNode(rpcUsers = listOf(user)).getOrThrow()
                 val nodeName = nodeHandle.nodeInfo.singleIdentity().name
@@ -108,63 +105,6 @@ class NodeStatePersistenceTests {
 fun isQuasarAgentSpecified(): Boolean {
     val jvmArgs = ManagementFactory.getRuntimeMXBean().inputArguments
     return jvmArgs.any { it.startsWith("-javaagent:") && it.endsWith("quasar.jar") }
-}
-
-@CordaSerializable
-data class Message(val value: String)
-
-data class MessageState(val message: Message, val by: Party, override val linearId: UniqueIdentifier = UniqueIdentifier()) : LinearState, QueryableState {
-    override val participants: List<AbstractParty> = listOf(by)
-
-    override fun generateMappedObject(schema: MappedSchema): PersistentState {
-        return when (schema) {
-            is MessageSchemaV1 -> MessageSchemaV1.PersistentMessage(
-                    by = by.name.toString(),
-                    value = message.value
-            )
-            else -> throw IllegalArgumentException("Unrecognised schema $schema")
-        }
-    }
-
-    override fun supportedSchemas(): Iterable<MappedSchema> = listOf(MessageSchemaV1)
-}
-
-object MessageSchema
-object MessageSchemaV1 : MappedSchema(
-        schemaFamily = MessageSchema.javaClass,
-        version = 1,
-        mappedTypes = listOf(PersistentMessage::class.java)) {
-
-    @Entity
-    @Table(name = "messages")
-    class PersistentMessage(
-            @Column(name = "by")
-            var by: String,
-
-            @Column(name = "value")
-            var value: String
-    ) : PersistentState()
-}
-
-const val MESSAGE_CONTRACT_PROGRAM_ID = "net.corda.test.node.MessageContract"
-
-open class MessageContract : Contract {
-    override fun verify(tx: LedgerTransaction) {
-        val command = tx.commands.requireSingleCommand<Commands.Send>()
-        requireThat {
-            // Generic constraints around the IOU transaction.
-            "No inputs should be consumed when sending a message." using (tx.inputs.isEmpty())
-            "Only one output state should be created." using (tx.outputs.size == 1)
-            val out = tx.outputsOfType<MessageState>().single()
-            "Message sender must sign." using (command.signers.containsAll(out.participants.map { it.owningKey }))
-
-            "Message value must not be empty." using (out.message.value.isNotBlank())
-        }
-    }
-
-    interface Commands : CommandData {
-        class Send : Commands
-    }
 }
 
 @StartableByRPC
